have lib/xmi just convert xmi to midi files, use separate utility for dumping to a file
